requirements.txt
azure-ai-projects
azure-identity
openai
semgrep
PyGithub>=2.1
google-genai
tree_sitter>=0.22
neo4j>=5.12